[feat] TDD 기반 user api 구현

+++ src/users/users.service.ts
@@ -1,11 +1,11 @@
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { User } from './entities/users.entity';
 
 @Injectable()
-export class UsersService {
+export class UserService {
   constructor(
     @InjectRepository(User)
     private userRepository: Repository<User>,
@@ -30,7 +30,10 @@ export class UsersService {
   }
 
   async removeUser(id: number) {
+    const user = await this.userRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException('user not found');
+    }
     await this.userRepository.delete(id);
-    return id;
   }
 }
